+++ frontend/components/common/ControlledTestArea.tsx
