+++ frontend/components/common/utils/createTakeAwayOrder.ts
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
 
-export async function  processOrder({
+export async function  CreateTakeAwayOrder({
   data,
   orderList,
   totalPrice,
